pyproject.toml
[project]
name = "easy-gemini-agent-engine"
version = "0.1.0"
description = "Simple Gemini agent with built-in tools for Agent Engine deployment"
license = {text = "Apache-2.0"}
requires-python = ">=3.11,<3.14"
dependencies = [
    "google-adk>=1.15.0,<2.0.0",
    "google-cloud-aiplatform[agent-engines]>=1.128.0,<2.0.0",
    "google-cloud-secret-manager>=2.16.0,<3.0.0",
    "nest-asyncio>=1.5.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "google-api-python-client>=2.100.0,<3.0.0",
    "google-auth>=2.20.0,<3.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.23.8,<1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gemini_agent"]

[tool.uv]
override-dependencies = []

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple/"
default = true

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
